+++ src/Cli/Cli.ts
@@ -5,8 +5,7 @@ import yargs from 'yargs-parser';
 import { CliArgs, parseArgv } from '../argv-parser';
 import { CliCommand, CliCommandInstance, createCliCommand, getCliCommand } from '../cli-command';
 import { log } from '../log';
-import { HelpPresenter, LogPresenter, VersionPresenter } from '../presenter';
-import { DisplayLevel } from '../presenter/Display';
+import { DisplayLevel, HelpPresenter, LogPresenter, VersionPresenter } from '../presenter';
 import { buildContext } from './CliContext';
 import { CliContext } from './interfaces';
 import { loadConfig } from './loadConfig';
