+++ src/presenter/index.ts
@@ -1,5 +1,3 @@
-export * from './Display';
 export * from './interfaces';
 export * from './PlainPresenter';
 export * from './plainPresenterFactory';
-export * from './presenter';
