+++ src/presenter/plainPresenterFactory.ts
@@ -1,6 +1,6 @@
 
-import { PresenterOption, LogPresenter, HelpPresenter, VersionPresenter, Inquirer } from './presenter'
-import { PlainPresenter } from './PlainPresenter'
+import { HelpPresenter, Inquirer, LogPresenter, PresenterOption, VersionPresenter } from './interfaces';
+import { PlainPresenter } from './PlainPresenter';
 
 export const plainPresenterFactory = {
   createCliPresenter(options: PresenterOption): LogPresenter & HelpPresenter & VersionPresenter {
